+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-  - NuvlaBox - set tags in commissioning
+  - NuvlaBox - set tags on commission operation
+  - Voucher - add country mappings
+  - Voucher - add correlation between voucher and voucher-discipline
   - NuvlaBox - make tags attribute editable for users
   - NuvlaBox - make location attribute editable for users
   - Voucher discipline - new resource